Rationalize file paths & adopting to macFUSE 4
This is based on Dave Griffin’s fixes, but changing the tool’s name from “afpfs-ng-OSX” to “afpfs-ng-mac”. This is to keep the differentiation from the original afpfs-ng, mainly because I changed its command to “mount_afp2”. But to also reflect Apple’s move from “OS X” to “macOS” and “FUSE for macOS” to “macFUSE” seven weeks ago.

Also, I keep the folder with “Mount AFP2.x volume.scpt” and “Unmount ALL AFP2.x volumes.scpt” in /Library/Scripts/ so they are also available from the Script Menu when logged in as a different user.

Fixes to meet the amended file locations of macFUSE v.4.x.